+++ apps/core_admin/urls.py
@@ -7,7 +7,9 @@ from . import views
 
 urlpatterns = [
     path('importar/', views.import_prospects, name='import_prospects'),
+    path('atendimentos/desativacao/', views.desativar_atendimentos_ixc, name='desativar_atendimentos_ixc'),
     path('restaurar-backup/', views.restore_backup, name='restore_backup'),
     path('teste-smtp/', views.smtp_test, name='smtp_test'),
     path('modelo/', views.download_template, name='download_template'),
+    path('atendimentos/desativacao/modelo/', views.download_template_desativacao_atendimento, name='download_template_desativacao_atendimento'),
 ]